update user conver to ldap
I have had so much help from this team I wanted to offer a submission for an AXL issue I ran into.  I am not a pro and still learning but I attempted to match the layout already in use.  Thanks again for all the help and maybe this is a little something to give back.